requirements.txt
asgiref==3.8.1
boto3==1.35.3
botocore==1.35.3
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.3
idna==3.7
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
phonenumbers==8.13.37
pillow==10.3.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
razorpay==1.4.2
requests==2.32.2
s3transfer==0.10.2
setuptools==71.1.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
